feat(docs): update Class 'Producto', upload enums in different Classes
