+++ backend/tiendita/src/main/java/com/bootcamp/model/enums/DisponibilidadProducto.java
@@ -1,0 +1,16 @@
+package com.bootcamp.model.enums;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+@JsonFormat(shape = JsonFormat.Shape.STRING)
+public enum DisponibilidadProducto {
+    DISPONIBLE,
+    VENDIDO,
+    PAUSADO;
+
+    @JsonCreator
+    public static DisponibilidadProducto fromString(String value) {
+        return DisponibilidadProducto.valueOf(value.toUpperCase());
+    }
+}
